Parte do Login feita

+++ backend/src/main/java/br/com/docvet/docvet/service/TipoAnimalService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.docvet.docvet.domain.TipoAnimal;
+import br.com.docvet.docvet.domain.error.NotFoundException;
 import br.com.docvet.docvet.repository.TipoAnimalRepository;
 
 @Service
@@ -15,7 +16,7 @@ public class TipoAnimalService {
     private TipoAnimalRepository repository;
 
     public TipoAnimal getOne(Integer id){
-        return repository.findById(id).get();
+        return repository.findById(id).orElseThrow(()-> new NotFoundException("Tipo de animal não encontrado: id="+id));
     }
 
     public List<TipoAnimal> getAll(){
@@ -27,14 +28,21 @@ public class TipoAnimalService {
     }
 
     public void update(Integer id, TipoAnimal tipoAnimal){
-        TipoAnimal tipoExistente = repository.findById(id).get();
+        if (!repository.existsById(tipoAnimal.getId())) {
+            throw new NotFoundException("Tipo animal não encontrado: id="+tipoAnimal.getId());
+        }
+        // TipoAnimal tipoExistente = repository.findById(id).get();
 
-        tipoExistente.setNome(tipoAnimal.getNome());
+        // tipoExistente.setNome(tipoAnimal.getNome());
 
-        repository.saveAndFlush(tipoExistente);
+        repository.saveAndFlush(tipoAnimal);
     }
 
     public void delete(Integer id){
+        if (!repository.existsById(id)) {
+            throw new NotFoundException("Tipo animal não encontrado: id="+id);
+
+        }
         repository.deleteById(id);
     }
